Split some stats files by job.
Job chain optimizations for additive workloads avoid sending old datums
to workers. In order for these old files to make it into a job's output
commit, the new file trees must be merged with the pre-existing files in
the parent commit. This works correctly for successful jobs, because
they only ever add stats files which are completely new.

Once a job fails, there are two issues. First, any subsequent
interaction with a failed datum will generate stats, which will be
appended to existing files. We can avoid having to separate the appended
data and match it to jobs by writing different jobs' stats to different
files in the first place.

Second, datums which are skipped because they were processed correctly
by a previous, failed job still need their stats in the final commit.
If a job is additive, these files are present in the parent commit
already, and so don't need to be remerged. Thus we can fix this by
avoiding including stats from skipped datums for additive jobs.

An alternative solution is to change merge behavior for the stats
filetrees. This would guarantee consistent results without changing
the stats commit format. However, rebuilding the stats hashtree
accounting for overwrites would be more expensive.